+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.27.3
+## 2.28.0
 
 *2025-06-18*
 
@@ -11,6 +11,11 @@
 **Fixed**
 
 - The CodeMirror editors in the preference dialog are now localized when used on a MediaWiki site
+
+**Changed**
+
+- [WikiEditor](./mw/README.md#wikieditor) toolbar buttons are now hidden if the textarea element is read-only on a MediaWiki site
+- The [wikiEditor](./mw/README.md#wikieditor) extension now also adds buttons to a pre-existing toolbar on a MediaWiki site
 
 ## 2.27.2